refactor name handler-broaker : path
